syz-manager: restore pprof handlers
Commit a72e2c3992 switched to custom mux in order to add compression.
But that broke the default http/pprof handlers since they are registered
only in the default mux.
We could manually register http/pprof handlers in our mux,
but this looks like lots of code, not a single call.
So restore the use of the default mux. CompressHandler can perfectly
work with it and does not require a custom mux.
Also add a layer of indirection to make any such changes easier in future.