syz-ci: always upload build assets
The existing rebuild check is not enough anyway -- as a result, we get
build records in the DB that do not have any assets (e.g. the ones that
appeared after syz-ci updated itself).

Instead do the upload unconditionally, now that we calculate hashes
first and don't re-upload asset with the same hash, it's not a big
problem.