prog: add an attribute for syscalls which should not be minimized
Create a `no_minimize` attribute to be used with syscalls that
`syzkaller` should not try to modify when minimizing a program that
produces a bug. The intention is to use this with syscalls that are
expensive to minimize, such as `syz_mount_image`.

Currently there are no `no_minimize` syscalls, but the next commit will
add some.