dashboard: always close a dev_appserver instance
Currently, if one of the ctx.Close asserts fails, we never close the
connection to the dev_appserver instance we use.

As a result, the application created by `go test` is left with lots of
orphaned dev_appserver children and `go test` endlessly waits for them
to stop.

Defer the aetest.Instance.Close() call to prevent this from happening.