pkg/build: include all build output in FreeBSD buildkernel errors
Continue digging into syzbot's FreeBSD build failures: the buildkernel
step looks like it's failing because it's using the host compiler, which
is too old and doesn't support KASAN.  We *should* be populating the
objdir with a toolchain in the earlier make step, though.  So let's
include both kernel-toolchain and buildkernel output in error messages
and see if that helps expose what's happening.