dashboard: integrate bug reminders and discussions
Add one more condition for skipping bugs from reminders -- time of the
last user reply to bug-related discussions. Let it be 2 weeks by
default.

At the same time, decrease the default value for the minimum bug
reporting date. If we reported a bug a week ago and nobody has reacted,
it still makes sense to include the bug in the report.