Revert "dashboard: bump CONFIG_FRAME_WARN for Android kernels"
This reverts commit ca8ef56e86a43b2bb65cff557215b94f1efc354e.

The bump of CONFIG_FRAME_WARN was done for Android 5.10 with the hope
to get rid of the -Wframe-larger-than errors that were hit when
KASAN_STACK && CC_IS_CLANG were set. Unfortunately this didn't fix the
build as the stack instrumentation when using clang is broken and the
errors were hit again as we can't predict the increase of the stack. The
fix is to disable CONFIG_WERROR which now is done because we just bumped
the Android 5.10 kernel config (which contains CONFIG_WERROR=y) and the
kernel config tweak that disables CONFIG_WERROR now does its job
(see dashboard/config/linux/bits/base.yml).

Signed-off-by: Tudor Ambarus <tudor.ambarus@linaro.org>