pkg/cover, syz-manager: support coverage filtering with modules
Apply module conversion to filter bitmap to support coverage
filtering when modules are used.

Coverage bitmap is decanonicalized to match each instance's
module PCs.

This was done by decanonicalizing the map of bitmap PCs before bitmap
creation to simplify conversion. I could convert the final bitmap or
pass in the canonicalization instance to createCoverageFilter() to
prevent this change.

The "coverFilter" used by the manager is only used on
canonicalized PCs, so it remains unchanged.