pkg/vcs: force git branch checkout
We are currently seeing errors that there are "untracked working
tree files which would be overrwritten by checkout".

This error occurs when files of the same path differ between branches,
regardless of whether the repository is clean or not.

Forcing FETCH_HEAD checkout, then moving repair() to after checkout
to clean repository to the current checkout.

Additionally, added a second "force" to git clean to clean directories
with '.git' subdirectories.